+++ CHANGELOG.md
@@ -5,7 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added 
+- PII scan rule to detect aadhar card and PAN card information
 
+### Changed 
+- Fixed false positives arising in PII Scan rule
+- Detect '-' separated numbers along with space separated (credit card number or identity card numbers can be '-' separated too.)
 
 ## [1.0.0] - 2023-07-24
 ### Added